requirements.txt
paho-mqtt==1.6.1 
# If on Windows use 1.6.1 because of compatibility issues
